requirements.txt
pydantic>=2.10.4
pandas
requests

logfire>=2.11.1
openai==1.58.1
reportlab==4.2.5

python-dotenv

scrapybara==2.3.0
playwright>=1.49.1

httpx==0.27.0
pymongo

anthropic

tldextract